Liballoc try 1

+++ include/kmalloc.h
@@ -3,15 +3,15 @@
 
 #include <stdint.h>
 
-uint32_t kmalloc_real(uint32_t size, int align, uint32_t * physical);
+uintptr_t kmalloc_real(uint32_t size, int align, uintptr_t * physical);
 
-uint32_t kmalloc_a(uint32_t size);
+uintptr_t kmalloc_a(uint32_t size);
 
-uint32_t kmalloc_p(uint32_t size, uint32_t * physical);
+uintptr_t kmalloc_p(uint32_t size, uintptr_t * physical);
 
-uint32_t kmalloc_ap(uint32_t size, uint32_t * physical);
+uintptr_t kmalloc_ap(uint32_t size, uintptr_t * physical);
 
-uint32_t kmalloc(uint32_t size);
+uintptr_t kmalloc(uint32_t size);
 
 void kfree(void * p);
